backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
firebase-admin
supabase
neo4j
pinecone-client

langchain
langchain-community
langchain-core
langchain-openai
langchain-pinecone
langchain-huggingface
sentence-transformers

docling
docling-core
pandas
pillow
beautifulsoup4
markdownify